requirements.txt
numpy==1.21.0
icecream==2.1.1
torchvision==1.13.1
